Add client filter unit tests
